# backend/openapi.json
{
  "components": {
    "schemas": {
      "AcceptedResponse": {
        "properties": {
          "status": {
            "const": "check_email",
            "default": "check_email",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "AcceptedResponse",
        "type": "object"
      },
      "CodeRequest": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 20,
            "minLength": 6,
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "CodeRequest",
        "type": "object"
      },
...
      "ConsentPurpose": {
        "description": "Separate purposes (docs/spec/10; REQ-CON-01).",
        "enum": [
          "marketing",
          "reminders",
          "whatsapp",
          "profiling",
          "github_import",
          "tier2_llm_assistant",
          "tier2_llm_moderation"
        ],
        "title": "ConsentPurpose",
        "type": "string"
      },
...
      "EmailRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "EmailRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "OrgKind": {
        "description": "Axis 1, Org Type (docs/spec/03).",
        "enum": [
          "company",
          "sme",
          "sacco_mfi",
          "university_tvet",
          "school",
          "national_govt",
          "county_govt",
          "ngo_pbo",
          "development_partner"
        ],
        "title": "OrgKind",
        "type": "string"
      },
...
      "OrgSignup": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/OrgKind"
          },
          "legal_name": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Legal Name",
            "type": "string"
          }
        },
        "required": [
          "legal_name",
          "kind"
        ],
        "title": "OrgSignup",
        "type": "object"
      },
...
      "RecoveryCodesResponse": {
        "properties": {
          "recovery_codes": {
            "items": {
              "type": "string"
            },
            "title": "Recovery Codes",
            "type": "array"
          }
        },
        "required": [
          "recovery_codes"
        ],
        "title": "RecoveryCodesResponse",
        "type": "object"
      },
...
      "SessionResponse": {
        "properties": {
          "mfa_required": {
            "title": "Mfa Required",
            "type": "boolean"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "required": [
          "user",
          "mfa_required"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
      "SignupRequest": {
        "additionalProperties": false,
        "properties": {
          "accept_terms": {
            "title": "Accept Terms",
            "type": "boolean"
          },
          "consents": {
            "additionalProperties": {
              "type": "boolean"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/ConsentPurpose"
            },
            "title": "Consents",
            "type": "object"
          },
          "display_name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "locale": {
            "default": "en",
            "enum": [
              "en",
              "sw"
            ],
            "title": "Locale",
            "type": "string"
          },
          "org": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrgSignup"
              },
              {
                "type": "null"
              }
            ]
          },
          "password": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "side": {
            "enum": [
              "developer",
              "org"
            ],
            "title": "Side",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "display_name",
          "side",
          "accept_terms"
        ],
        "title": "SignupRequest",
        "type": "object"
      },
      "StaffRole": {
        "enum": [
          "admin",
          "moderator",
          "support"
        ],
        "title": "StaffRole",
        "type": "string"
      },
      "TokenRequest": {
        "additionalProperties": false,
        "properties": {
          "token": {
            "maxLength": 200,
            "minLength": 20,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "TokenRequest",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "email_verified": {
            "title": "Email Verified",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "locale": {
            "title": "Locale",
            "type": "string"
          },
          "staff_role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StaffRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "totp_enabled": {
            "title": "Totp Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "display_name",
          "locale",
          "email_verified",
          "staff_role",
          "totp_enabled"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Bridge (working name) API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/magic-link": {
      "post": {
        "description": "Email a sign-in link (or a fresh verification link to an unverified account). Always 202.",
        "operationId": "magic_link_api_auth_magic_link_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Magic Link",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/magic-link/consume": {
      "post": {
        "operationId": "consume_api_auth_magic_link_consume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Consume",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/mfa/verify": {
      "post": {
        "operationId": "mfa_verify_api_auth_mfa_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mfa Verify",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/signup": {
      "post": {
        "operationId": "signup_api_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signup",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/step-up": {
      "post": {
        "operationId": "step_up_api_auth_step_up_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Step Up",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/totp/confirm": {
      "post": {
        "operationId": "totp_confirm_api_auth_totp_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Totp Confirm",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/verify-email/resend": {
      "post": {
        "description": "Same behaviour as /magic-link: unverified accounts get a verification link. Always 202.",
        "operationId": "resend_verification_api_auth_verify_email_resend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resend Verification",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}